requirements.txt
Flask>=2.0.0,<3.0.0
requests==2.31.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
langchain==0.2.15
langchain-community==0.2.15
langchain-huggingface>=0.0.3
sqlalchemy==2.0.19
transformers==4.30.2
torch==2.4.0
sentencepiece==0.1.99
accelerate==0.20.3
huggingface_hub==0.19.0